requirements.txt
astroid==2.4.2
awsebcli==3.19.3
backports.entry-points-selectable==1.1.0
botocore==1.19.43
cement==2.8.2
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
cycler==0.10.0
defusedxml==0.6.0
distlib==0.3.2
dj-database-url==0.5.0
Django==3.0.3
django-allauth==0.43.0
django-ckeditor==5.9.0
django-cors-headers==3.3.0
django-extensions==3.1.5
django-heroku==0.3.1
django-js-asset==1.2.2
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-rest-knox==4.1.0
django-rest-passwordreset==1.2.1
django-templated-mail==1.1.1
django-timezone-field==4.1.1
djangorestframework==3.11.0
djongo==1.3.6
djoser==2.1.0
dnspython==2.2.0
filelock==3.0.12
future==0.16.0
gunicorn==20.1.0
idna==2.10
isort==5.6.4
itypes==1.2.0
Jinja2==3.0.3
jmespath==0.10.0
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
MarkupSafe==2.0.1
matplotlib==3.4.2
mccabe==0.6.1
mongoengine==0.23.1
ndg-httpsclient==0.5.1
numpy==1.21.1
oauthlib==3.1.0
pathspec==0.5.9
Pillow==8.0.1
pipenv==2021.11.23
platformdirs==2.2.0
psycopg2==2.8.6
pusher==3.2.0
pyasn1==0.4.8
pycparser==2.20
PyJWT==2.3.0
pylint==2.6.0
pymongo==3.12.3
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.4
python3-openid==3.2.0
pytz==2019.3
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
semantic-version==2.5.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.2.4
stripe==2.64.0
termcolor==1.1.0
toml==0.10.2
typed-ast==1.4.2
uritemplate==4.1.1
urllib3==1.25.11
virtualenv==20.7.0
virtualenv-clone==0.5.7
wcwidth==0.1.9
whitenoise==6.0.0
wrapt==1.12.1
